requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.3
django-import-export==3.2.0
django-jazzmin==2.6.0
et-xmlfile==1.1.0
import-export==0.3.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.0.0
pip==23.1.2
PyYAML==6.0
setuptools==67.8.0
sqlparse==0.4.4
tablib==3.5.0
tzdata==2023.3
wheel==0.40.0
xlrd==2.0.1
xlwt==1.3.0
